Split out api loaders and models
