+++ api/loaders/reaction.js
@@ -1,18 +1,9 @@
 // @flow
-import {
-  getReactions,
-  getThreadReactions,
-  getReactionsByIds,
-} from '../models/reaction';
+import { getReactions, getReactionsByIds } from '../models/reaction';
 import createLoader from './create-loader';
 
 export const __createReactionLoader = createLoader(
   messageIds => getReactions(messageIds),
-  'group'
-);
-
-export const __createThreadReactionLoader = createLoader(
-  threadIds => getThreadReactions(threadIds),
   'group'
 );
